+++ test/observer_impl_test.rb
@@ -9,11 +9,15 @@ class ObserverTest < Minitest::Test
   CLASS      = Observer
   ONE        = 1
   EMPTY_HASH = {}
+  NILCLASS_I = nil
+  TEST_SYMBOL = :test_symbol
+  NODE1 = Node.new(NILCLASS_I, TEST_SYMBOL, NILCLASS_I)
 
   # test_conf_doc_f_ex().
   # @description
   #   The .travis.yml, CODE_OF_CONDUCT.md, Gemfile, LICENSE.txt, README.md,
-  #   and .yardopts files exist.
+  #   .yardopts, .gitignore, Changelog.md, CODE_OF_CONDUCT.md,
+  #   observer_impl.gemspec, Gemfile.lock, and Rakefile files exist.
   def test_conf_doc_f_ex()
 
     assert_path_exists('.travis.yml')
@@ -22,6 +26,12 @@ class ObserverTest < Minitest::Test
     assert_path_exists('LICENSE.txt')
     assert_path_exists('README.md')
     assert_path_exists('.yardopts')
+    assert_path_exists('.gitignore')
+    assert_path_exists('Changelog.md')
+    assert_path_exists('CODE_OF_CONDUCT.md')
+    assert_path_exists('observer_impl.gemspec')
+    assert_path_exists('Gemfile.lock')
+    assert_path_exists('Rakefile')
 
   end
 
@@ -74,11 +84,7 @@ class ObserverTest < Minitest::Test
   # @description
   #   An unobservable instance.
   def test_cas_x1()
-
-    assert_raises(ArgumentError) {
-      Observer.add_subject(EMPTY_HASH)
-    }
-
+    assert_raises(ArgumentError) { Observer.add_subject(EMPTY_HASH) }
   end
 
   # test_cas_x2().
@@ -86,11 +92,10 @@ class ObserverTest < Minitest::Test
   #   An observable instance.
   def test_cas_x2()
 
-    n = Minitest::Mock.new()
-    n.expect(:class, Node)
-    assert_raises(NameError) {
-      Observer.add_subject(n)
-    }
+    r = Observer.add_subject(NODE1)
+    no_singleton = NodeObserver.instance()
+    assert_operator(no_singleton, 'subject', NODE1)
+    assert_nil(r)
 
   end
 
@@ -109,67 +114,76 @@ class ObserverTest < Minitest::Test
 
   # test_crs_x2().
   # @description
-  #   An observable instance.
+  #   The argument is an observable object and a subject.
   def test_crs_x2()
 
-    n = Minitest::Mock.new()
-    n.expect(:class, Node)
-    assert_raises(NameError) {
-      Observer.remove_subject(n)
+    Observer.add_subject(NODE1)
+    r = Observer.remove_subject(NODE1)
+    assert_nil(r)
+    no_singleton = NodeObserver.instance()
+    refute_operator(no_singleton, 'subject', NODE1)
+
+  end
+
+  # test_crs_x3().
+  # @description
+  #   The argument is an observable object, and not a subject.
+  def test_crs_x3()
+
+    assert_raises(ArgumentError, "#{NODE1} is not a subject.") {
+      Observer.remove_subject(NODE1)
     }
 
   end
 
-  # Observer.notify(pub_i = nil).
+  # Observer.notify(instance = nil).
 
   # test_notify_x1().
   # @description
-  #   A Node subject.
+  #   The argument is not an observable instance.
   def test_notify_x1()
 
-    n = Minitest::Mock.new()
-    n.expect(:class, Node)
-    n.expect(:class, Node)
-    assert_raises(NameError) {
-      Observer.notify(n)
+    assert_raises(ArgumentError,
+                  "#{TEST_SYMBOL} is not an observable instance.") {
+      Observer.notify(TEST_SYMBOL)
     }
 
-  end
-
-  # test_notify_x2().
-  # @description
-  #   A subject excluding a Node.
-  def test_notify_x2()
-    assert_nil(Observer.notify(nil))
   end
 
   # Observer.changed(instance = nil).
 
   # test_changed_x1().
   # @description
-  #   A Node.
+  #   Any object excluding observable objects.
   def test_changed_x1()
-
-    n = Minitest::Mock.new()
-    n.expect(:class, Node)
-    n.expect(:class, Node)
-    assert_raises(NameError) {
-      Observer.changed(n)
-    }
-
+    refute_operator(Observer, 'changed', TEST_SYMBOL)
   end
 
   # test_changed_x2().
   # @description
-  #   The argument's class is not Node.
+  #   An observable object, and a subject.
   def test_changed_x2()
-    refute_operator(Observer, 'changed', nil)
+    Observer.add_subject(NODE1)
+    refute_operator(Observer, 'changed', NODE1)
+  end
+
+  # test_changed_x3().
+  # @description
+  #   An observable object, and not a subject.
+  def test_changed_x3()
+    refute_operator(Observer, 'changed', NODE1)
   end
 
   # teardown().
   # @description
-  #  Cleanup.
+  #   Cleanup.
   def teardown()
+
+    no_singleton = NodeObserver.instance()
+    if (no_singleton.subject(NODE1))
+      no_singleton.remove(NODE1)
+    end
+
   end
 
 end